requirements.txt
Pillow>=10.0
Flask>=3.0
PyYAML>=6.0
requests>=2.31
webdavclient3>=3.14
